+++ polaris-react/src/components/RadioButton/tests/RadioButton.test.tsx
@@ -76,28 +76,6 @@ describe('<RadioButton />', () => {
     });
   });
 
-  describe('Focus className', () => {
-    it('on keyUp adds a keyFocused class to the input', () => {
-      const radioButton = mountWithApp(<RadioButton label="Radio" />);
-
-      radioButton.find('input')!.trigger('onKeyUp');
-      expect(radioButton).toContainReactComponent('input', {
-        className: 'Input keyFocused',
-      });
-    });
-
-    it('on change does not add a keyFocused class to the input', () => {
-      const radioButton = mountWithApp(<RadioButton label="Radio" />);
-      const radioInput = radioButton.find('input');
-      radioInput!.trigger('onChange', {
-        currentTarget: radioInput!.domNode as HTMLInputElement,
-      });
-      expect(radioButton).not.toContainReactComponent('input', {
-        className: 'Input keyFocused',
-      });
-    });
-  });
-
   describe('onBlur()', () => {
     it('is called when the input is focused', () => {
       const spy = jest.fn();