+++ polaris-react/src/components/Tooltip/Tooltip.tsx
@@ -1,11 +1,21 @@
 import React, {useEffect, useState, useRef, useCallback} from 'react';
+import type {
+  ShapeBorderRadiusScale,
+  SpacingSpaceScale,
+} from '@shopify/polaris-tokens';
 
 import {Portal} from '../Portal';
 import {findFirstFocusableNode} from '../../utilities/focus';
 import {useUniqueId} from '../../utilities/unique-id';
 import {useToggle} from '../../utilities/use-toggle';
+import {classNames} from '../../utilities/css';
 
 import {TooltipOverlay, TooltipOverlayProps} from './components';
+import styles from './Tooltip.scss';
+
+export type Width = 'default' | 'wide';
+export type Padding = 'default' | Extract<SpacingSpaceScale, '4'>;
+export type BorderRadius = Extract<ShapeBorderRadiusScale, '1' | '2'>;
 
 export interface TooltipProps {
   /** The element that will activate to tooltip */
@@ -30,6 +40,23 @@ export interface TooltipProps {
   activatorWrapper?: string;
   /** Visually hidden text for screen readers */
   accessibilityLabel?: string;
+  /**
+   * Width of content
+   * @default 'default'
+   */
+  width?: Width;
+  /**
+   * Padding of content
+   * @default 'default'
+   */
+  padding?: Padding;
+  /**
+   * Border radius of the tooltip
+   * @default '1'
+   */
+  borderRadius?: BorderRadius;
+  /** Override on the default z-index of 400 */
+  zIndexOverride?: number;
   /* Callback fired when the tooltip is activated */
   onOpen?(): void;
   /* Callback fired when the tooltip is dismissed */
@@ -45,6 +72,10 @@ export function Tooltip({
   preferredPosition = 'below',
   activatorWrapper = 'span',
   accessibilityLabel,
+  width = 'default',
+  padding = 'default',
+  borderRadius = '1',
+  zIndexOverride,
   onOpen,
   onClose,
 }: TooltipProps) {
@@ -101,11 +132,19 @@ export function Tooltip({
         accessibilityLabel={accessibilityLabel}
         onClose={noop}
         preventInteraction={dismissOnMouseOut}
+        width={width}
+        padding={padding}
+        borderRadius={borderRadius}
+        zIndexOverride={zIndexOverride}
       >
         {content}
       </TooltipOverlay>
     </Portal>
   ) : null;
+
+  const wrapperClassNames = classNames(
+    activatorWrapper === 'div' && styles.TooltipContainer,
+  );
 
   return (
     <WrapperComponent
@@ -121,6 +160,7 @@ export function Tooltip({
       onMouseOver={handleMouseEnterFix}
       ref={setActivator}
       onKeyUp={handleKeyUp}
+      className={wrapperClassNames}
     >
       {children}
       {portal}